chore(playbook): refresh resume hint
